feat(core): migrate retry memory runner test to jdbc/kafka runner test
