+++ jdbc-h2/src/test/java/io/kestra/runner/h2/H2RunnerRetryTest.java
@@ -1,0 +1,7 @@
+package io.kestra.runner.h2;
+
+import io.kestra.jdbc.runner.JdbcRunnerRetryTest;
+
+public class H2RunnerRetryTest extends JdbcRunnerRetryTest {
+
+}
