+++ jdbc-mysql/src/test/java/io/kestra/runner/mysql/MysqlRunnerRetryTest.java
@@ -1,0 +1,7 @@
+package io.kestra.runner.mysql;
+
+import io.kestra.jdbc.runner.JdbcRunnerRetryTest;
+
+public class MysqlRunnerRetryTest extends JdbcRunnerRetryTest {
+
+}
