+++ jdbc-postgres/src/test/java/io/kestra/runner/postgres/PostgresRunnerRetryTest.java
@@ -1,0 +1,7 @@
+package io.kestra.runner.postgres;
+
+import io.kestra.jdbc.runner.JdbcRunnerRetryTest;
+
+public class PostgresRunnerRetryTest extends JdbcRunnerRetryTest {
+
+}
